requirements.txt
pandas
scipy
fastparquet
pytorch
hyperopt
numpy
scikit-learn
